added functions, fixed bugged, patched Wu's code

+++ together/removeStudentData.py
@@ -4,6 +4,13 @@ from PyQt4 import QtCore
 from removeStudentDialog import Ui_removeStudentDialog
 from PyQt4.QtSql import *
 
+'''
+class removeStudentData -
+# contains the functions and querys
+# needed to remove a student and
+# all other data linked to the students
+# if there are no other links
+'''
 class removeStudentData(QtGui.QDialog):
     def __init__(self):
         QtGui.QDialog.__init__(self)
@@ -21,6 +28,7 @@ class removeStudentData(QtGui.QDialog):
             QtGui.QMessageBox.warning(
                 self, 'Error', 'database contecting error')
 
+        # get student name
         self.student_query = QSqlQuery()
         self.student_query.exec_("Select Student_name FROM Student")
         while self.student_query.next():
@@ -30,23 +38,23 @@ class removeStudentData(QtGui.QDialog):
             item.setText(self.name)
             self.model.appendRow(item)
 
+        # set up list view model
         self.removeStu.studentListView.setModel(self.model)
         self.removeStu.studentListView.setSelectionMode(QtGui.QAbstractItemView.ExtendedSelection)
-        
+
         self.removeStu.ok_btn.clicked.connect(self.submit)
         self.removeStu.cancel_btn.clicked.connect(self.close)
 
     def submit(self):
-        studentList = []
-        addressCounter = 0
-        guardCounter = 0
+        studentList = [] #selected student
+        stuAddressCounter = 0 # counter to see if address used more then once
+        guardCounter = 0 # same for guardian
         secondaryGuardCounter = 0
-        #self.removeStu.studentListView.model().removeRow(refresh)
 
+        # build student
         for selected in self.removeStu.studentListView.selectedIndexes():
             studentList.append(str(selected.data()))
         for student in studentList:
-            print(student)
             getId = QSqlQuery()
             getId.exec_("Select Student_id FROM Student WHERE \
                          Student_name = '%s'" % student)
@@ -54,15 +62,18 @@ class removeStudentData(QtGui.QDialog):
                 record = getId.record()
                 self.id = int(record.value(0))
 
+            # get addresses
             getAddress = QSqlQuery()
+            # select address for a single student and then check if others have same address 
             getAddress.exec_("Select Student_address from Student WHERE \
                 Student_address = (Select Student_address from Student WHERE\
                 Student_name = '%s')" % student)
             while getAddress.next():
                 record = getAddress.record()
                 addressId = int(record.value(0))
-                addressCounter += 1
+                stuAddressCounter += 1
 
+            # get guardians
             getGuardian = QSqlQuery()
             getGuardian.exec_("Select Guardian_primary, Guardian_secondary from Student WHERE \
                 Guardian_primary = (Select Guardian_primary from Student WHERE \
@@ -72,13 +83,11 @@ class removeStudentData(QtGui.QDialog):
                 record = getGuardian.record()
                 guardId = int(record.value(0))
                 guardIdSecondary = int(record.value(1))
-                print(guardId)
-                print(guardIdSecondary)
                 guardCounter += 1
                 if (guardIdSecondary != None):
                     secondaryGuardCounter += 1
             
-
+            # double check with user before removal 
             self.confirmMessage = "Are you sure you want to remove '%s' from the system? This will remove all relevant information, including payments and billing. This action can not be reversed." \
                                  %(student)
             self.confirmReply = QtGui.QMessageBox.question(self, 'Confirm', 
@@ -91,49 +100,74 @@ class removeStudentData(QtGui.QDialog):
                 self.confirm.exec_("DELETE from Credits WHERE Student_id = %d" % self.id)
                 self.confirm.exec_("DELETE from Discount WHERE Student_id = %d" % self.id)
                 self.confirm.exec_("DELETE from Payment WHERE Student_id = %d" % self.id)
-            
-                if (addressCounter == 1):
-                    self.confirm.exec_("DELETE from Address WHERE Address_id = %d" % addressId)
+
+                '''
+                # delete if this student is the only student that lives at their address.
+                # Prevent delete if someone still in the system lives at the address
+                # and make sure not to delete the NONE addreess
+                '''
                     
-                if (guardCounter == 1):
-                    addressCounter = 0
+                '''
+                # delete if this guardian is the only student that lives at their address.
+                # Prevent delete if someone still in the system lives at the address
+                # and make sure not to delete the NONE addreess
+                '''
+                if (guardCounter == 1 and guardId != 0):
+                    guardAddressCounter = 0
+                    # check guardian addresses
                     getAddress.exec_("Select Guardian_address from Guardian WHERE \
                        Guardian_address = (Select Guardian_address from Guardian WHERE\
                        Graudian_id = %d)" % guardId)
                     while getAddress.next():
                         record = getAddress.record()
-                        addressId = int(record.value(0))
-                        print(addressId)
-                        addressCounter += 1
+                        guardAddressId = int(record.value(0))
+                        if (guardAddressId != None):
+                            guardAddressCounter += 1
 
-                    if (addressCounter == 1):
-                        self.confirm.exec_("DELETE from Address WHERE Address_id = %d" % addressId)
-                        
+                    
+                    # Delete guardian if only one student tied to them    
                     self.confirm.exec_("DELETE from Guardian WHERE Guardian_id = %d" % guardId)
 
-                if (secondaryGuardCounter == 1):
-                    addressCounter = 0
+                if (secondaryGuardCounter == 1 and guardIdSecondary != 0):
+                    # same for secondary guardian
+                    secondaryAddressCounter = 0
                     getAddress.exec_("Select Guardian_address from Guardian WHERE \
                        Guardian_address = (Select Guardian_address from Guardian WHERE\
                        Graudian_id = %d)" % guardIdSecondary)
                     
                     while getAddress.next():
                         record = getAddress.record()
-                        addressId = int(record.value(0))
-                        addressCounter += 1
-
-                    if (addressCounter == 1):
-                        self.confirm.exec_("DELETE from Address WHERE Address_id = %d" % addressId)
-
+                        secondaryAddressId = int(record.value(0))
+                        if (secondaryAddressId != None):
+                            secondaryAddressCounter += 1
                     
                     self.confirm.exec_("DELETE from Guardian WHERE Guardian_id = %d" % guardIdSecondary)
 
+                    
+                '''
+                Here are the three checks to see if the address can be deleted safely
+                1. if 1 student and 1 guardian and 1 secondary
+                2. 1 student and 1 guardian
+                3. 1 student
+                Otherwise do not delete address and finsh removal
+                '''
+
+                if ((stuAddressCounter == 1 and guardAddressCounter == 1 and secondaryAddressCounter == 1) and addressId != 0):
+                    self.confirm.exec_("DELETE from Address WHERE Address_id = %d" % addressId)
+                elif ((stuAddressCounter == 1 and guardAddressCounter == 1 and secondaryAddressCounter == 0) and addressId != 0):
+                    self.confirm.exec_("DELETE from Address WHERE Address_id = %d" % addressId)
+                elif ((stuAddressCounter == 1 and guardAddressCounter == 0 and secondaryAddressCounter == 0) and addressId != 0):
+                    self.confirm.exec_("DELETE from Address WHERE Address_id = %d" % addressId)
                 self.message = "Student removed"
+                #refresh list
                 self.refreshModel()
             
                 self.confirmReply = QtGui.QMessageBox.information(self, 'Message', 
                     self.message)
 
+    '''
+    refreshModel resets the list view after a removal
+    '''
     def refreshModel(self):
         self.model.clear()
         self.student_query = QSqlQuery()
